fix(sftp): align transfer concurrency defaults
