+++ application/state/sftp/transferConcurrency.test.ts
@@ -12,9 +12,9 @@ import {
   runSftpTransferWorkers,
 } from "./transferConcurrency";
 
-test("defaults folder file transfers to two concurrent files", () => {
+test("defaults folder file transfers to six concurrent files", () => {
   assert.equal(resolveSftpTransferConcurrency(() => null), DEFAULT_SFTP_FILE_TRANSFER_CONCURRENCY);
-  assert.equal(DEFAULT_SFTP_FILE_TRANSFER_CONCURRENCY, 2);
+  assert.equal(DEFAULT_SFTP_FILE_TRANSFER_CONCURRENCY, 6);
 });
 
 test("defaults directory listing fanout to four concurrent readdirs", () => {
@@ -51,18 +51,18 @@ test("keeps directory listing concurrency within the supported range", () => {
   );
 });
 
-test("limits default multi-file transfer scheduling to two concurrent workers", async () => {
+test("limits default multi-file transfer scheduling to six concurrent workers", async () => {
   let active = 0;
   let maxActive = 0;
 
-  await runSftpTransferWorkers([1, 2, 3, 4], () => null, async () => {
+  await runSftpTransferWorkers([1, 2, 3, 4, 5, 6, 7], () => null, async () => {
     active += 1;
     maxActive = Math.max(maxActive, active);
     await new Promise((resolve) => setTimeout(resolve, 5));
     active -= 1;
   });
 
-  assert.equal(maxActive, 2);
+  assert.equal(maxActive, 6);
 });
 
 test("runBoundedConcurrency respects an explicit limit", async () => {
